refactor: TableGroup <-> OrderTable 양방향 해결
