+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -2,15 +2,18 @@ package kitchenpos.application;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 
 import kitchenpos.application.dto.request.OrderTableRequest;
 import kitchenpos.application.dto.request.TableGroupRequest;
 import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.OrderTables;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
@@ -19,13 +22,11 @@ import kitchenpos.repository.TableGroupRepository;
 @Service
 public class TableGroupService {
 
-    private final OrderRepository orderRepository;
     private final OrderTableRepository orderTableRepository;
     private final TableGroupRepository tableGroupRepository;
 
-    public TableGroupService(OrderRepository orderRepository, OrderTableRepository orderTableRepository,
+    public TableGroupService(OrderTableRepository orderTableRepository,
                              TableGroupRepository tableGroupRepository) {
-        this.orderRepository = orderRepository;
         this.orderTableRepository = orderTableRepository;
         this.tableGroupRepository = tableGroupRepository;
     }
@@ -33,8 +34,12 @@ public class TableGroupService {
     @Transactional
     public TableGroup create(final TableGroupRequest tableGroupRequest) {
         List<OrderTable> orderTables = toOrderTables(tableGroupRequest.getOrderTables());
+        validateMoreThanTwoOrderTable(orderTables);
+        validateIsEmptyAndNotContainTableGroup(orderTables);
         validateNotDuplicatedOrNotExist(orderTables, tableGroupRequest.getOrderTables());
-        TableGroup tableGroup = new TableGroup(LocalDateTime.now(), orderTables);
+        TableGroup tableGroup = new TableGroup(LocalDateTime.now());
+
+        orderTables.forEach(orderTable -> orderTable.registerTableGroup(tableGroup));
 
         return tableGroupRepository.save(tableGroup);
     }
@@ -57,12 +62,32 @@ public class TableGroupService {
     public void ungroup(final Long tableGroupId) {
         final TableGroup tableGroup = tableGroupRepository.findById(tableGroupId)
                 .orElseThrow(IllegalArgumentException::new);
-        final List<OrderTable> orderTables = tableGroup.getOrderTables();;
+        final OrderTables orderTables = new OrderTables(
+                        orderTableRepository.findAllByTableGroupId(tableGroup.getId())
+        );
 
-        if (tableGroup.hasOrderTableWhichStatusIsCookingOrMeal()) {
+        if (orderTables.hasOrderTableWhichStatusIsCookingOrMeal()) {
             throw new IllegalArgumentException();
         }
 
-        orderTables.forEach(tableGroup::deleteOrderTable);
+        orderTables.unregisterTableGroup();
+    }
+
+    private void validateMoreThanTwoOrderTable(List<OrderTable> orderTables) {
+        if (CollectionUtils.isEmpty(orderTables) || orderTables.size() < 2) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    private void validateIsEmptyAndNotContainTableGroup(List<OrderTable> orderTables) {
+        for (final OrderTable orderTable : orderTables) {
+            validateOrderTableIsEmptyAndNotContainTableGroup(orderTable);
+        }
+    }
+
+    private void validateOrderTableIsEmptyAndNotContainTableGroup(OrderTable orderTable) {
+        if (!orderTable.isEmpty() || Objects.nonNull(orderTable.getTableGroup())) {
+            throw new IllegalArgumentException();
+        }
     }
 }
